% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_VaRTest.R
\name{multilevel_VaRTest}
\alias{multilevel_VaRTest}
\title{Multilevel conditional coverage tests.}
\usage{
  multilevel_VaRTest(alphas, actual, VaR, confidence, m = 5, B = 2000)
}
\arguments{
  \item{alphas}{vector of quantiles}
  
  \item{actual}{vector of actual returns}
  
  \item{VaR}{matrix of VaR forecasts. Order needs to be the same as in alphas.}
  
  \item{confidence}{confidence level for unconditional coverage tests}
  
  \item{m}{number of lags for multilevel Pearson's chi-square test. Leccadito et. al recommend a lag of 5.}
    
    \item{B}{number of Monte-Carlo simulations for Pearson's chi-square test.}
    }
\value{
  a list with the following items:
    \describe{
      \item{"likelihood ratio CC"}{Christoffersen conditional coverage test LR statistic.}
      \item{"critical value CC"}{Christoffersen conditional coverage test critical value.}
      \item{"p-value CC"}{Christoffersen conditional coverage test p-value.}
      \item{"decision CC"}{Christoffersen conditional coverage test decision on H0 given the confidence level.}
      \item{"likelihood ratio UC"}{Christoffersen unconditional coverage test LR statistic.}
      \item{"critical value UC"}{Christoffersen unconditional coverage test critical value.}
      \item{"p-value UC"}{Christoffersen unconditional coverage test p-value.}
      \item{"decision UC"}{Christoffersen unconditional coverage test decision on H0 given the confidence level.}
      \item{"X_m"}{Pearson's chi-square test statistic}
        \item{"p-value pearson"}{Pearson's chi-square test p-value}
      \item{"decision.pearson"}{Pearson's chi-square test decision on H0 of given the confidence level.}}
      }
        \description{
        This function implements multilevel unconditional coverage tests of VaR forecasts for different quantiles. These tests achieve a higher statistical power then multiple unilevel tests. \cr
        Theory based on "Evaluating the accuracy of value-at- risk forecasts: New multilevel tests" by Leccadito, Bofelli & Urga (2014). \cr
        Code for the multilevel Christoffersen UC test is based on the function "VaRTest" from the "rugarch" package by Alexios Ghalanos.
        }